+++ src/lib/components/editor/URLEditor.svelte
@@ -1,5 +1,8 @@
 <script lang="ts">
+    import Icon from './Icon.svelte';
     import TextEditor from './TextEditor.svelte';
+    import LinkIcon from './icons/link.svg?raw';
+    import Link from '../app/Link.svelte';
 
     export let url: string;
     export let validator: (text: string) => undefined | string;
@@ -15,3 +18,4 @@
     clip
     save={(url) => edit(url)}
 />
+<Link to={url}>↗</Link>